Handle the like button function
So when user already like, it will remove the like instead. Still don't know how to change the label of the like button to be unlike when user click on the button, so now it stays as like. But you can see in terminal if it is adding a reaction or removing a reaction